+++ potato_client/package.json
@@ -8,7 +8,8 @@
     "react-linkify": "^0.2.1",
     "react-markdown": "^2.5.0",
     "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1"
+    "react-router-dom": "^4.1.1",
+    "react-scrollable-anchor": "^0.4.2"
   },
   "devDependencies": {
     "react-scripts": "1.0.7"
