+++ potato_client/src/App.css
@@ -104,6 +104,14 @@ body {
   font-family: 'PT Sans', sans-serif;
 }
 
+.titlebock {
+  margin: 1px 50px;
+  font-size: 16px;
+  font-weight: bold;
+  text-align: justify;
+  font-family: 'Cormorant SC', serif;
+}
+
 .hotel {
   line-height: 10%;
 }
